refactored cpu module

+++ dlgrad/builder/float_utils_build.py
@@ -10,7 +10,7 @@ cdef = """
 void neg(float *x, float *out, int numel);
 void cexp(float *x, float *out, int numel);
 void clog(float *x, float *out, int numel);
-void cpow(float *x, float *out, int val, int numel);
+void cpow(float *x, float *out, float val, int numel);
 void csqrt(float *x, float *out, int numel);
 """
 ffi.cdef(cdef)
